add values and operators

+++ main.c
@@ -3,8 +3,6 @@
 #include <string.h>
 
 #include "common.h"
-#include "chunk.h"
-#include "debug.h"
 #include "vm.h"
 
 static void repl() {
@@ -33,7 +31,7 @@ static char* readFile(const char* path) {
     }
 
     fseek(file, 0L, SEEK_END);
-    size_t fileSize = ftell(file);
+    size_t fileSize = (size_t) ftell(file);
     rewind(file);
 
     char* buffer = (char*)malloc(fileSize + 1);
